Optimize pr-str by making more use of StringBuffer.
Previously, a huge number of temporary vectors and lazy sequences were
built up and concat-ed, which ran reasonably on newer JS engines (e.g.
V8) but suffered from miserable performance on old engines (such as
IE8).  By passing a StringBuffer down the callstack and appending to
that, performance gains of 10-100X can be seen in older JS engines.

For backwards compatibility, the new fastpath uses a new IPrintWith
protocol, and the old IPrintable is left in place.  The old protocol
is used as a fallback for items that do not satisfy IPrintWith.